Update patch set 1
Patch Set 1:

I know what this is. There was a change in DLRN to allow having the same spec for CBS and DLRN builds, and in the case of Puppet modules it requires a change in the spec.

See https://review.rdoproject.org/r/gitweb?p=puppet/puppet-aodh-distgit.git;a=commitdiff;h=refs/heads/rpm-master for an example. TL;DR is that the %setup line needs to be:

%setup -q -n openstack-magnum-%{upstream_version}

Patch-set: 1
Reviewer: Gerrit User 113 <113@270e2033-b340-4cff-9539-693957ebf0e7>
Label: Code-Review=0